+++ packages/api/src/settings/transformers.ts
@@ -7,22 +7,30 @@ import { SettingDocument, SettingDto } from './types';
  * @returns {SettingDto}
  */
 export const settingTransformer = (setting: SettingDocument): SettingDto => {
-  const { _id, key, type, label, description, group, options } = setting;
-  let { value, valueRealized } = setting;
-  if (type === 'JSON' && value) {
-    value = JSON.stringify(JSON.parse(value), null, 2);
-    valueRealized = JSON.stringify(JSON.parse(value), null, 2);
+  const { _id, key, type, label, description, group, subgroup, options } =
+    setting;
+  let { defaultValue, value, valueRealized } = setting;
+  if (type === 'JSON') {
+    if (defaultValue) {
+      defaultValue = JSON.stringify(JSON.parse(defaultValue), null, 2);
+    }
+    if (value) {
+      value = JSON.stringify(JSON.parse(value), null, 2);
+      valueRealized = JSON.stringify(JSON.parse(value), null, 2);
+    }
   }
   return {
     _id,
     key,
+    label,
+    description,
+    defaultValue,
     value,
     valueRealized,
     type,
-    label,
-    description,
-    group,
     options,
+    group,
+    subgroup,
   };
 };