+++ packages/api/src/tests/settings.spec.ts
@@ -33,10 +33,12 @@ describe('GET /api/settings/all', () => {
       'key',
       'label',
       'type',
-      'group',
       'description',
+      'defaultValue',
       'value',
-      'valueRealized'
+      'valueRealized',
+      'group',
+      'subgroup'
     );
   });
 
@@ -75,7 +77,8 @@ describe('GET /api/settings/:id', () => {
       'group',
       'description',
       'value',
-      'valueRealized'
+      'valueRealized',
+      'defaultValue'
     );
   });
 
@@ -142,7 +145,8 @@ describe('PATCH /api/admin/settings/:id/set', () => {
       'type',
       'label',
       'description',
-      'group'
+      'group',
+      'defaultValue'
     );
   });