+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **API:** Transformer option to include/exclude header row in csv exports #614 #622
+- **Frontend:** Most settings have now been given default values. To reset a setting to its default value, click the "Reset to default" button that appears next to the setting header. #580 #613
 
 ### Changed
 
 - **Discord-bot**: Updated the bot to use the new `discord.js` version 14 #619
 - **Discord-bot**: Minor refactors to make code easier to read #619
+- **Frontend:** The settings pages have been reorganised slightly and given section descriptions with links to documentation. #586 #631
 
 ### Fixed